Install backends to /usr/libexec instead of /usr/bin (which is in $PATH)
(note: this requires #456 to be merged first. Splitted so that this
topic can be discussed :))

IIUIC the backends are not supposed to be executed directly by the user,
but spawned using spectrum2_manager. In this case, the binaries should
not be in $PATH, i.e outside of /usr/bin.

This patch installs the backends to /usr/libexec [1], as this matches the
FSH intention for this directory.

[1] https://refspecs.linuxfoundation.org/FHS_3.0/fhs/ch04s07.html